% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{md.data}
\alias{md.data}
\title{md.data}
\usage{
md.data(data, randomsample = FALSE)
}
\arguments{
\item{data}{data.frame}

\item{randomsample}{r}
}
\description{
Creates information of covariates that are copies of covariates from an existing data set.
This function call must be added to the \code{\link{md.simparams}} object.
}
\examples{

\dontrun{
library(missDeaths)

sim = md.simparams() +
   md.data(data) 
}
}

