% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{md.dateuniform}
\alias{md.dateuniform}
\title{md.dateuniform}
\usage{
md.dateuniform(name, start, end)
}
\arguments{
\item{name}{name of the covariate}

\item{start, end}{lower and upper limits of the date interval}
}
\description{
Creates information of a uniformly distributed date covariate with a specified lower and upper limits.
This function call must be added to the \code{\link{md.simparams}} object.
}
\examples{

\dontrun{
library(missDeaths)

sim = md.simparams() +
   md.dateuniform("birth", as.Date("1930-1-1"), as.Date("1960-1-1"))
}

}

