% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{md.expprep}
\alias{md.expprep}
\title{Calculates individual expected survival beyond the observation time}
\usage{
md.expprep(D, time)
}
\arguments{
\item{D}{Demographic information}

\item{time}{Observation time}
}
\description{
Calculates individual expected survival beyond the observation time
}
\keyword{internal}
