% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{md.impute}
\alias{md.impute}
\title{Correctly impute missing information of possible deaths using population mortality}
\usage{
md.impute(data, f, maxtime, D, ratetable, iterations = 4)
}
\arguments{
\item{data}{a data.frame in which to interpret the variables named in the
formula.}

\item{f}{a formula object, with the response on the left of a ~ operator,
and the terms on the right. The response must be a survival object as
returned by the \code{Surv} function.}

\item{maxtime}{maximum potential observation time (number of days).

where \code{status}=0 equals \code{time}.

where \code{status}=1 equals potential time of right censoring if no event
would be observed.}

\item{D}{demographic information compatible with \code{md.survcox}, \code{md.impute} 
and \code{md.survnp}, see \code{\link{md.D}}.}

\item{ratetable}{a population mortality table, default is \code{slopop}}

\item{iterations}{the number of iteration steps to be performed, default is
4}
}
\value{
an array of times with imputed times of death that can be used instead of the
unavailable complete data set to get unbiased estimates, ie. in \code{\link[survival]{coxph}}.
}
\description{
An iterative approach is used in this method to estimate the conditional
distribution required to correctly impute the times of deaths using
population mortality tables.\cr\cr
  Note, that simply imputing expected survival times may seem intuitive, 
but does not give unbiased estimates, since the right censored individuals 
are not a random subsample of the patients.
}
\examples{
library(missDeaths)
data(slopop)

data(observed)
observed$time = observed$time*365.2425
D = md.D(age=observed$age*365.2425, sex=observed$sex, year=(observed$year - 1970)*365.2425)
newtimes = md.impute(observed, Surv(time, status) ~ age + sex + iq + elevation, 
  observed$maxtime*365.2425, D, slopop, iterations=4)

#Cumulative incidence function
cif = survfit(Surv(observed$time, observed$status)~1)
cif$surv = 1 - cif$surv
cif$upper = 1 - cif$upper
cif$lower = 1 - cif$lower
plot(cif)

#Net survival (NOTE: std error is slightly underestimated!)
surv.net = survfit(Surv(newtimes, observed$status)~1)
summary(surv.net, times=c(3,9)*365.2425)
plot(surv.net)

#Event free survival (NOTE: std error is slightly underestimated!)
surv.efs = survfit(Surv(newtimes, 1 * (observed$status | (newtimes != observed$time)))~1)
summary(surv.efs, times=c(3,9)*365.2425)
plot(surv.efs)

}
\references{
Stupnik T., Pohar Perme M. (2015) "Analysing disease recurrence
with missing at risk information." Statistics in Medicine 35. p1130-43.
\url{https://onlinelibrary.wiley.com/doi/abs/10.1002/sim.6766}
}
\seealso{
\code{\link{md.survcox}}
}
