% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{md.sample}
\alias{md.sample}
\title{md.sample}
\usage{
md.sample(name, array, weights=NULL)
}
\arguments{
\item{name}{name of the covariate}

\item{array}{vector of elements from which to choose from}

\item{weights}{vector of probability weights for obtaining the elements of the vector being sampled or NULL if all values have equal probabilities}
}
\description{
Creates information of a covariate that represents a random sample (with replacement) of the provided values.
This function call must be added to the \code{\link{md.simparams}} object.
}
\examples{

\dontrun{
library(missDeaths)

sim = md.simparams() +
   md.sample("X", c(0, 1, 2), c(0.2, 0.3, 0.5)) 
}
}
