% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{md.sex}
\alias{md.sex}
\title{md.sex}
\usage{
md.sex(name, maleperc = 0.5, asfactor = FALSE)
}
\arguments{
\item{name}{name of the covariate}

\item{maleperc}{percentage of males (value = 1) versus females (value = 2)}

\item{asfactor}{specifies whether the resulting sex covariate is factorized using as.factor or not}
}
\description{
Creates information of a sex covariate that is a special case of Bernoulli covariate
specifying 1 for male and 2 for female sex. 
This function call must be added to the \code{\link{md.simparams}} object.
}
\examples{

\dontrun{
library(missDeaths)

sim = md.simparams() +
   md.sex("sex", 0.5)
}

}
