% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{md.survprob}
\alias{md.survprob}
\title{Initializes the missDeaths population table cache}
\usage{
md.survprob(year, age, time, sex)
}
\arguments{
\item{year}{year}

\item{age}{age}

\item{time}{time}

\item{sex}{sex}
}
\description{
Initializes the missDeaths population table cache
}
\keyword{internal}
