\name{observed}
\docType{data}
\alias{observed}
\title{Simulated data}
\description{
  This data set with missing information on possible deaths can be used to test the missDeaths library functions.
}
\usage{observed}
\format{A \code{data.frame} containing 10000 observations.}
\keyword{datasets}